Protect link by detaching it from DOM
Previously, the link was neutered by wiping href and text contents.
To restore, it's outerHTML would be reset to it's original value.

Now, detach the link from the DOM and swap in a dummy placeholder link
in it's place. Hopefully this results in fewer overall changes to the DOM.
Also, it should be more future-resilient if changes are made to the site,
as the original link's event listeners are preserved.